mcp/postgres/requirements.txt
mcp
psycopg2-binary